Oversized commit
This commit is the result of many experiments. It's a complete refactor
to the previous implementation. This one is completely working with the
backend.

Everything is now a singletone, you don't instatiate the
headquarters-node, you just initialize it. I don't believe it's
necessary to have multiple instances of an object that connects to the
same api.

It's also using es6 now, there is a small gulpfile to compile and to run
the tests.