Add client credentials flow
All the endpoint tests use this new authentication instead of
authorization code, since most of the times this requests will be made
from a bot, where the token doesn't identify a person.

Added a bot demo, that uses the client credentials flow. The interface
is really similar.

Updated jshintrc to better support the project.